requirements.txt
pytorch 
torchvision 
torchaudio
pytest



